requirements-dev.txt
google-cloud-storage
google-auth
google-api-python-client
requests
crcmod
